fix: move other end points into invoice.js
